Task 6

+++ 0x00-hello_world/6-size.c
@@ -14,10 +14,10 @@ int main(void)
 	long long int d;
 	float e;
 
-	printf("size of a char: %lu byte(s)\n", (long unsigned)sizeof(a));
-	printf("size of an int: %lu byte(s)\n", (long unsigned)sizeof(b));
-	printf("size of a long int: %lu byte(s)\n", (long unsigned)sizeof(c));
-	printf("size of a long long int: %lu byte(s)\n", (long unsigned)sizeof(d));
-	printf("size of a float: %lu byte(s)\n", (long unsigned)sizeof(e));
+	printf("size of a char: %d byte(s)\n", sizeof(a));
+	printf("size of an int: %d byte(s)\n", sizeof(b));
+	printf("size of a long int: %d byte(s)\n", sizeof(c));
+	printf("size of a long long int: %d byte(s)\n", sizeof(d));
+	printf("size of a float: %d byte(s)\n", sizeof(e));
 	return (0);
 }
